adding trotterized time evolution
